backend/requirements.txt

fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-huggingface
huggingface_hub
firecrawl
pymupdf
pandas
